+++ packages/react-redux/src/slices/cartoSlice.d.ts
@@ -1,7 +1,7 @@
 import { Credentials, SourceProps } from '@carto/react-api/';
 import { FiltersLogicalOperators, Viewport, _FilterTypes } from '@carto/react-core';
 import { CartoBasemapsNames, GMapsBasemapsNames } from '@carto/react-basemaps/';
-import { InitialCartoState, CartoState, ViewState } from '../types';
+import { InitialCartoState, CartoState, ViewState, Source, FilterValues } from '../types';
 import { AnyAction, Reducer } from 'redux';
 import { Feature, Polygon, MultiPolygon } from 'geojson';