+++ packages/react-redux/src/types.d.ts
@@ -1,7 +1,7 @@
-import { Credentials } from '@carto/react-api';
+import { Credentials, SourceProps } from '@carto/react-api';
 import { OauthApp } from '@carto/react-auth';
 import { CartoBasemapsNames, GMapsBasemaps } from '@carto/react-basemaps';
-import { Viewport } from '@carto/react-core';
+import { Viewport, FiltersLogicalOperators, _FilterTypes } from '@carto/react-core';
 import { Geometry } from 'geojson';
 
 export type ViewState = {
@@ -14,6 +14,19 @@ export type ViewState = {
   width?: number,
   height?: number,
 }
+
+type FilterValues = string[] | number[] | number[][]
+
+export type SourceFilters = {
+  [column: string]: Partial<Record<_FilterTypes, { values: FilterValues; owner?: string; params?: Record<string, unknown>; }>>
+}
+
+export type Source = SourceProps & {
+  id: string;
+  filters?: SourceFilters;
+  filtersLogicalOperator?: FiltersLogicalOperators;
+  isDroppingFeatures?: boolean;
+};
 
 type InitialCarto2State = {
   accountsUrl?: string,
@@ -45,7 +58,7 @@ export type CartoState = {
   geocoderResult: Record<string,any> | null,
   error: null, // TODO: remove from state?
   layers: { [key: string]: string },
-  dataSources: { [key: string]: string },
+  dataSources: { [key: string]: Source },
   spatialFilter: Geometry,
   featuresReady: { [key: string]: boolean },
   featureSelectionEnabled: boolean,