+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Not released
 
 - Freeze widgets during time-series animation [#804](https://github.com/CartoDB/carto-react/pull/804)
+- Added filterable prop to TimeSeriesWidget [#808](https://github.com/CartoDB/carto-react/pull/808)
+- Fix dataSources store type [#807](https://github.com/CartoDB/carto-react/pull/807)
 
 ## 2.3